+++ frontend/frontend.ts
@@ -1,4 +1,4 @@
-import {DocumentUpdatedMessage, MessageToUi, MessageToUiType, ScanResultDto} from "../shared/types";
+import {DocumentUpdatedMessage, MessageToUi, MessageToUiType, SettingsUpdatedMessage} from "../shared/types";
 import {ComponentInfoElement} from "./src/elements/components-tab/component-info.element";
 import {Util} from "../shared/util";
 import {ComponentsListHeaderElement} from "./src/elements/components-tab/components-list-header.element";
@@ -10,6 +10,7 @@ import {SearchInputElement} from "./src/elements/common/search-input.element";
 import {InstancesGroupElement} from "./src/elements/instances-tab/instances-group.element";
 import {ResizeCornerElement} from "./src/elements/common/resize-corner.element";
 import {ScanResultsManager} from "./src/scan-results-manager";
+import {SettingsBarElement} from "./src/elements/common/settings-bar.element";
 
 ComponentInfoElement.register();
 ComponentsListHeaderElement.register();
@@ -19,6 +20,7 @@ InstanceInfoElement.register();
 SearchInputElement.register();
 InstancesGroupElement.register();
 ResizeCornerElement.register();
+SettingsBarElement.register();
 
 export const util = new Util("CM-Frontend");
 
@@ -28,6 +30,12 @@ onmessage = (event) => {
     if (messageToUi.type === MessageToUiType.DOCUMENT_UPDATE) {
         const typedMessage = messageToUi as DocumentUpdatedMessage;
         ScanResultsManager.getInstance().handleDocumentUpdate(typedMessage.payload);
+    } else if (messageToUi.type === MessageToUiType.SETTINGS_UPDATE) {
+        const typedMessage = messageToUi as SettingsUpdatedMessage;
+        const settingsBar = document.querySelector('app-settings-bar') as SettingsBarElement;
+        if (settingsBar) {
+            settingsBar.setAutoRefreshState(typedMessage.payload.autoRefresh);
+        }
     }
 }